fix(library): WIP add connections implementation
