+++ src/message/message.ts
@@ -1,5 +1,6 @@
 import { Envelope, EmptyEnvelope } from './envelope';
 import * as Types from '../types';
+import * as Model from '../model';
 import * as Mobx from 'mobx';
 
 export enum MessageType {
@@ -240,6 +241,7 @@ export type ConnectPatternLibraryResponse = Envelope<
 	MessageType.ConnectPatternLibraryResponse,
 	{
 		analysis: Types.LibraryAnalysis;
+		library: Model.PatternLibrary;
 		path: string;
 		previousLibraryId: string | undefined;
 	}
