+++ src/renderer/message-handlers/connect-pattern-library.ts
@@ -1,7 +1,6 @@
 import * as M from '../../message';
 import { MessageHandlerContext, MessageHandler } from '../create-handlers';
 import * as Types from '../../types';
-import * as Model from '../../model';
 import * as uuid from 'uuid';
 
 export function connectPatternLibrary({
@@ -14,31 +13,21 @@ export function connectPatternLibrary({
 			return;
 		}
 
-		const analysis = m.payload.analysis;
+		const id = m.payload.previousLibraryId;
+		if (!id) {
+			return;
+		}
 
-		const library = Model.PatternLibrary.create({
-			id: uuid.v4(),
-			name: analysis.name,
-			version: analysis.version,
-			origin: Types.PatternLibraryOrigin.UserProvided,
-			patternProperties: [],
-			patterns: [],
-			bundle: analysis.bundle,
-			bundleId: analysis.id,
-			description: analysis.description,
-			state: Types.PatternLibraryState.Connected
-		});
-
-		library.import(analysis, { project });
-
-		project.addPatternLibrary(library);
+		console.log(m.payload.previousLibraryId);
+		// library.import(analysis, { project });
+		// project.addPatternLibrary(library);
 		store.getApp().setRightSidebarTab(Types.RightSidebarTab.ProjectSettings);
 		store.commit();
 
 		store.getSender().send({
 			id: uuid.v4(),
 			payload: {
-				id: library.getId(),
+				id,
 				path: m.payload.path
 			},
 			type: M.MessageType.ConnectedPatternLibraryNotification
